update setup.py to use setup.cfg
